requirements.txt
# Core dependencies
numpy>=1.20.0
requests>=2.25.0
pydantic>=1.8.0

# Audio processing
pyaudio>=0.2.11
sounddevice>=0.4.4
soundfile>=0.10.3

# STT - faster-whisper
faster-whisper>=0.5.0
ffmpeg-python>=0.2.0

# TTS - Coqui TTS
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.25.0
accelerate>=0.20.0
TTS>=0.17.0

# LLM - Ollama
ollama>=0.1.0

# Memory and embeddings
sentence-transformers>=2.2.0
chromadb>=0.4.0
# sqlite3 is part of the Python standard library

# Utilities
python-dotenv>=0.19.0
tqdm>=4.62.0
colorlog>=6.7.0
pyyaml>=6.0.0
PySimpleGUI>=4.60.0
